example/gcc/WebLoad/requirements.txt
Flask==2.3.3
pyserial==3.5
Werkzeug==2.3.7
cryptography==41.0.7
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psutil==5.9.6